Changed a lot, separated kernels and devices
Devices can now use a range of compatible kernels.
I implemented a working "optimized" kernel (instead of naive), but
in reality it is actually much slower. As this shouldn't be the
case, it needs to be fixed. This is a heck of a lot easier to
debug and benchmark, however. As it stands, the naive kernel is
about 5x slower than BLAS for a 2k by 2k matrix, where the
"optimized" version is ~10x slower. So there is room to improve
and perhaps overtake BLAS.

The CPU works great with the naive kernel, however it is really
pretty slow. On small datasets it is actuall faster than the GPU,
however. Possible optimizations:
    - Zero copy path
    - CPU specific optimizations not yet investigated.